Load, extend, and stack all weeks of STEM rasters for a given result type

For one of four result types, loads all available weeks of rasters, extends them to the extent of the study (or to a custom extent), and stacks them into a RasterStack, with zeroes added (option to turn this off).

stack_stem(path, variable, st_extent = NA, add_zeroes = TRUE, weeks = NA,
  use_analysis_extent = TRUE)

Arguments

path
character; Full path to directory containing the STEM results for a single species.

variable
character; One of: 'abundance_ensemble_support', 'abundance_lower', 'abundance_upper', 'abundance_umean', and 'occurrence_umean'.

st_extent
list; Optional, use to limit the spatial Extent that the rasters are loaded into. Must set use_analysis_extent to FALSE.

add_zeroes
logical; Default is TRUE. Adds predicted and assumed zero values to the resulting layers. Set to FALSE to turn this behavior off.

weeks
numeric; Optional, a single integer or vector of integers from 1 to 52, representing the week(s) of the year to load. Defaults to all available. Note that if some weeks are not present, they will not be loaded.

use_analysis_extent
logical; Default is TRUE. If STEM results were run for a custom non-global extent, that extent object is stored in the configuration file. If TRUE, uses that analysis extent for loading the rasters. If FALSE, rasters are loaded to the full extent of the template_raster object.

Value

RasterStack containing all available weeks of result.

Examples

# NOT RUN {
sp_path <- "path to species STEM results"

raster_stack <- stack_stem(path = sp_path, variable = "abundance_umean")
# }