Template raster for extending the extents of STEM raster layers

A raster layer that is used to unify the extents of STEM results across weeks of the year. Currently, each individual STEM result layer (e.g., "abundance_umean") is stored by having all NA values trimmed, which results in raster layers that have different extents for each species for each week of the year. This template raster provides the full spatial extent of the study area.

template_raster

Format

A RasterLayer with pixels having value of 0 that represents the full spatial extent, the resolution, and the coordinate system of the STEM study extent.

Examples

str(template_raster)

#> Formal class 'RasterLayer' [package "raster"] with 12 slots
#>   ..@ file    :Formal class '.RasterFile' [package "raster"] with 13 slots
#>   .. .. ..@ name        : chr "/Users/mta45/Downloads/srd2016_wh_raster_template_3xagg.tif"
#>   .. .. ..@ datanotation: chr "FLT8S"
#>   .. .. ..@ byteorder   : chr "little"
#>   .. .. ..@ nodatavalue : num -Inf
#>   .. .. ..@ NAchanged   : logi FALSE
#>   .. .. ..@ nbands      : int 1
#>   .. .. ..@ bandorder   : chr "BIL"
#>   .. .. ..@ offset      : int 0
#>   .. .. ..@ toptobottom : logi TRUE
#>   .. .. ..@ blockrows   : int 1
#>   .. .. ..@ blockcols   : int 2267
#>   .. .. ..@ driver      : chr "gdal"
#>   .. .. ..@ open        : logi FALSE
#>   ..@ data    :Formal class '.SingleLayerData' [package "raster"] with 13 slots
#>   .. .. ..@ values    : logi(0) 
#>   .. .. ..@ offset    : num 0
#>   .. .. ..@ gain      : num 1
#>   .. .. ..@ inmemory  : logi FALSE
#>   .. .. ..@ fromdisk  : logi TRUE
#>   .. .. ..@ isfactor  : logi FALSE
#>   .. .. ..@ attributes: list()
#>   .. .. ..@ haveminmax: logi TRUE
#>   .. .. ..@ min       : num 0
#>   .. .. ..@ max       : num 0
#>   .. .. ..@ band      : int 1
#>   .. .. ..@ unit      : chr ""
#>   .. .. ..@ names     : chr "srd2016_wh_raster_template_3xagg"
#>   ..@ legend  :Formal class '.RasterLegend' [package "raster"] with 5 slots
#>   .. .. ..@ type      : chr(0) 
#>   .. .. ..@ values    : logi(0) 
#>   .. .. ..@ color     : logi(0) 
#>   .. .. ..@ names     : logi(0) 
#>   .. .. ..@ colortable: logi(0) 
#>   ..@ title   : chr(0) 
#>   ..@ extent  :Formal class 'Extent' [package "raster"] with 4 slots
#>   .. .. ..@ xmin: num -2e+07
#>   .. .. ..@ xmax: num -1109171
#>   .. .. ..@ ymin: num -6674483
#>   .. .. ..@ ymax: num 9229189
#>   ..@ rotated : logi FALSE
#>   ..@ rotation:Formal class '.Rotation' [package "raster"] with 2 slots
#>   .. .. ..@ geotrans: num(0) 
#>   .. .. ..@ transfun:function ()  
#>   ..@ ncols   : int 2267
#>   ..@ nrows   : int 1907
#>   ..@ crs     :Formal class 'CRS' [package "sp"] with 1 slot
#>   .. .. ..@ projargs: chr "+proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs"
#>   ..@ history : list()
#>   ..@ z       : list()

r <- raster()

#> Error in eval(expr, envir, enclos): could not find function "raster"

extended_raster <- extend(r, extent(template_raster))

#> Error in eval(expr, envir, enclos): could not find function "extend"